+++ src/ondusSenseGuard.ts
@@ -1,3 +1,4 @@
+import moment from 'moment';
 import { Service, PlatformAccessory } from 'homebridge';
 
 import { OndusPlatform } from './ondusPlatform';
@@ -69,17 +70,21 @@ export class OndusSenseGuard extends OndusAppliance {
     // set the Valve service characteristics
     this.valveService
       .setCharacteristic(this.ondusPlatform.Characteristic.Name, accessory.context.device.name)
+      .setCharacteristic(this.ondusPlatform.Characteristic.ValveType, this.ondusPlatform.Characteristic.ValveType.GENERIC_VALVE)
       .setCharacteristic(this.ondusPlatform.Characteristic.Active, this.ondusPlatform.Characteristic.Active.ACTIVE)
       .setCharacteristic(this.ondusPlatform.Characteristic.InUse, this.ondusPlatform.Characteristic.InUse.IN_USE)
-      .setCharacteristic(this.ondusPlatform.Characteristic.ValveType, this.ondusPlatform.Characteristic.ValveType.GENERIC_VALVE)
+      .setCharacteristic(this.ondusPlatform.Characteristic.StatusActive, this.ondusPlatform.Characteristic.Active.ACTIVE)
       .setCharacteristic(this.ondusPlatform.Characteristic.StatusFault, this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
       
     // register handlers for required characteristics of Valve service
     this.valveService.getCharacteristic(this.ondusPlatform.Characteristic.Active)
-      .on('get', this.handleActiveGet.bind(this))
-      .on('set', this.handleActiveSet.bind(this));
+      .on('get', this.handleValveServiceActiveGet.bind(this))
+      .on('set', this.handleValveServiceActiveSet.bind(this));
+  }
 
 
+  start() {
+    this.getHistoricalMeasurements();
   }
 
 
@@ -91,9 +96,38 @@ export class OndusSenseGuard extends OndusAppliance {
 
   // ---- CHARACTERISTICS HANDLER FUNCTIONS BELOW ----
 
+  setValveServiceActive(action: boolean) {
+    if (action) {
+      this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
+        this.ondusPlatform.Characteristic.Active.ACTIVE);
+    } else {
+      this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
+        this.ondusPlatform.Characteristic.Active.INACTIVE);
+    }
+  }
 
-  setValveServiceStatusFault(active: boolean) {
-    if (active) {
+  setValveServiceInUse(action: boolean) {
+    if (action) {
+      this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.InUse, 
+        this.ondusPlatform.Characteristic.InUse.IN_USE);
+    } else {
+      this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.InUse, 
+        this.ondusPlatform.Characteristic.InUse.NOT_IN_USE);
+    }
+  }
+
+  setValveServiceStatusActive(action: boolean) {
+    if (action) {
+      this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusActive, 
+        this.ondusPlatform.Characteristic.Active.ACTIVE);
+    } else {
+      this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusActive, 
+        this.ondusPlatform.Characteristic.Active.INACTIVE);
+    }
+  }
+
+  setValveServiceStatusFault(action: boolean) {
+    if (action) {
       this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.StatusFault, 
         this.ondusPlatform.Characteristic.StatusFault.GENERAL_FAULT);
     } else {
@@ -101,6 +135,8 @@ export class OndusSenseGuard extends OndusAppliance {
         this.ondusPlatform.Characteristic.StatusFault.NO_FAULT);
     }
   }
+
+  
 
 
   // ---- HTTP HANDLER FUNCTIONS BELOW ----
@@ -110,7 +146,7 @@ export class OndusSenseGuard extends OndusAppliance {
    */
   handleCurrentTemperatureGet(callback) {
     this.ondusPlatform.log.debug(`[${this.logPrefix}] Triggered GET CurrentTemperature`);
-    this.getMeasurements()
+    this.getLastMeasurements()
       .then ( () => {
         callback(null, this.currentTemperature);
       })
@@ -122,7 +158,7 @@ export class OndusSenseGuard extends OndusAppliance {
   /**
    * Handle requests to get the current value of the "Active" characteristic
    */
-  handleActiveGet(callback) {
+  handleValveServiceActiveGet(callback) {
     this.ondusPlatform.log.debug(`[${this.logPrefix}] Triggered GET Active`);
     //this.ondusPlatform.log.debug('GET reporting: ', this.currentValveState);
     this.getValveState()
@@ -138,7 +174,7 @@ export class OndusSenseGuard extends OndusAppliance {
   /**
    * Handle requests to set the "Active" characteristic
    */
-  handleActiveSet(value, callback) {
+  handleValveServiceActiveSet(value, callback) {
     this.ondusPlatform.log.debug(`[${this.logPrefix}] Triggered SET Active: ${value}`);
     
     if (!this.ondusPlatform.config['valve_control']) {
@@ -181,11 +217,52 @@ export class OndusSenseGuard extends OndusAppliance {
 
 
   /**
+   * Fetch Ondus Sense historical measurement data from the Ondus API. Returns a promise that will be resolved
+   * once measurement data has been queried from the Ondus API.
+  */
+  getHistoricalMeasurements() {
+    this.ondusPlatform.log.debug(`[${this.logPrefix}] Fetching historical temperature, flowrate, and pressure levels`);
+    // Fetch all registered appliance measurements
+    // If no measurements are present, the following is returned {"code":404,"message":"Not found"}
+
+    this.getApplianceMeasurements()
+      .then( measurement => {
+        //this.ondusPlatform.log.debug('res: ', measurement);
+        const measurementArray = measurement.body.data.measurement;
+        if (!Array.isArray(measurementArray)) {
+          this.ondusPlatform.log.debug(`[${this.logPrefix}] Unknown response ${measurement.body}`);
+        }
+        this.ondusPlatform.log.debug(`[${this.logPrefix}] Retrieved ${measurementArray.length} historical measurements`);
+        measurementArray.sort((a, b) => {
+          const a_ts = new Date(a.timestamp).getTime();
+          const b_ts = new Date(b.timestamp).getTime();
+          if(a_ts > b_ts) {
+            return 1;
+          } else if(a_ts < b_ts) {
+            return -1;
+          } else {
+            return 0;
+          }
+        });
+
+        // Add historical measurements to fakegato
+        measurementArray.forEach( value => {
+          this.historyService.addEntry({time: moment(value.timestamp).unix(), 
+            temp: value.temperature_guard});
+        });
+      })
+      .catch( err => {
+        this.ondusPlatform.log.error(`[${this.logPrefix}] Unable to retrieve historical temperature, flowrate, and pressure: ${err}`);
+      });
+  }
+
+
+  /**
   * Fetch Ondus Sense Guard measurement data. Returns a promise that will be resolved
    * once measurement data has been queried from the Ondus API.
   */
-  getMeasurements() {
-    this.ondusPlatform.log.debug(`[${this.logPrefix}] Updating temperature, flowrate, and pressure readings`);
+  getLastMeasurements() {
+    this.ondusPlatform.log.debug(`[${this.logPrefix}] Updating temperature, flowrate, and pressure levels`);
 
     // Fetch appliance measurements using current timestamp
     // If no measurements are present, the following is returned {"code":404,"message":"Not found"}
@@ -196,14 +273,14 @@ export class OndusSenseGuard extends OndusAppliance {
     // this promise when getApplianceCommand promise has been resolved and result processed.
     // This will prevent handleActiveGet() function returning incorrect value for currentValveState
     return new Promise<number>((resolve, reject) => {
-        
+      
       // Get fromDate measurements
       const fromDate = new Date(Date.now());
 
       // Retrieve latest Sense Guard measurement metrics
       this.getApplianceMeasurements(fromDate)
         .then( measurement => {
-        //this.ondusPlatform.log.debug('res: ', measurement);
+          //this.ondusPlatform.log.debug('res: ', measurement);
           const measurementArray = measurement.body.data.measurement;
           if (!Array.isArray(measurementArray)) {
             this.ondusPlatform.log.debug(`[${this.logPrefix}] Unknown response ${measurement.body}`);
@@ -233,22 +310,24 @@ export class OndusSenseGuard extends OndusAppliance {
           this.ondusPlatform.log.info(`[${this.logPrefix}] => Temperature: ${this.currentTemperature}˚C`);
 
           // Enable Active characteristics for temperature service
-          this.setTemperatureServiceActive(true);
-          
+          this.setTemperatureServiceStatusActive(true);
+        
           // Resolve promise to handleCurrentTemperatureGet()
           resolve(this.currentTemperature);
         })
         .catch( err => {
           this.ondusPlatform.log.error(`[${this.logPrefix}] Unable to update temperature: ${err}`);
-        
+      
           // Disable Active characteristics for temperature service
-          this.setTemperatureServiceActive(false);
-          
+          this.setTemperatureServiceStatusActive(false);
+        
           // Reject promise to handleCurrentTemperatureGet() and return last known temperature
           reject(this.currentTemperature);
         });
     });
   }
+
+
 
   /**
    * Retrieve Sense Guard valve state. Returns a promise that will be resolved
@@ -266,25 +345,23 @@ export class OndusSenseGuard extends OndusAppliance {
           this.currentValveState = res.body.command.valve_open;
           if (this.currentValveState) {
             this.ondusPlatform.log.debug(`[${this.logPrefix}] Main water inlet valve is open`);
-            this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
-              this.ondusPlatform.Characteristic.Active.ACTIVE);
-            this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.InUse, 
-              this.ondusPlatform.Characteristic.InUse.IN_USE);
+            this.setValveServiceActive(true);
+            this.setValveServiceInUse(true);
           } else {
             this.ondusPlatform.log.debug(`[${this.logPrefix}] Main water inlet valve is closed`);
-            this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
-              this.ondusPlatform.Characteristic.Active.INACTIVE);
-            this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.InUse, 
-              this.ondusPlatform.Characteristic.InUse.NOT_IN_USE);
+            this.setValveServiceInUse(false);
+            this.setValveServiceActive(false);
           }
 
           // Resolve promise to handleActiveGet()
+          this.setValveServiceStatusActive(true);
           resolve(this.currentValveState);
         })
         .catch(err => {
           this.ondusPlatform.log.error(`[${this.logPrefix}] Unable to retrieve main water inlet valve state: ${err}`);
 
           // Resolve promise to handleActiveGet() and return last known valve state
+          this.setValveServiceStatusActive(false);
           reject(this.currentValveState);
         });
     });
@@ -301,12 +378,10 @@ export class OndusSenseGuard extends OndusAppliance {
     // Log input
     if (valveState === OndusSenseGuard.VALVE_OPEN) {
       this.ondusPlatform.log.warn(`[${this.logPrefix}] Opening main water inlet valve`);
-      this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
-        this.ondusPlatform.Characteristic.Active.ACTIVE);
+      this.setValveServiceActive(true);
     } else {
       this.ondusPlatform.log.warn(`[${this.logPrefix}] Closing main water inlet valve`);
-      this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.Active, 
-        this.ondusPlatform.Characteristic.Active.INACTIVE);
+      this.setValveServiceActive(false);
     }
 
     // Construct payload for controlling valve state
@@ -323,16 +398,15 @@ export class OndusSenseGuard extends OndusAppliance {
           if (this.currentValveState) {
             this.ondusPlatform.log.warn(`[${this.logPrefix}] Main water inlet valve has been opened`);
             this.currentValveState = OndusSenseGuard.VALVE_OPEN;
-            this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.InUse, 
-              this.ondusPlatform.Characteristic.InUse.IN_USE);
+            this.setValveServiceInUse(true);
           } else {
             this.ondusPlatform.log.warn(`[${this.logPrefix}] Main water inlet valve has been closed`);
             this.currentValveState = OndusSenseGuard.VALVE_CLOSED;
-            this.valveService.updateCharacteristic(this.ondusPlatform.Characteristic.InUse, 
-              this.ondusPlatform.Characteristic.InUse.NOT_IN_USE);
+            this.setValveServiceInUse(false);
           }
 
           // Resolve promise to handleActiveSet()
+          this.setValveServiceStatusActive(true);
           resolve(this.currentValveState);
         })
         .catch(err => {
@@ -340,6 +414,7 @@ export class OndusSenseGuard extends OndusAppliance {
           this.ondusPlatform.log.error(`[${this.logPrefix}] Unable to retrieve valve state: ${err}`);
 
           // Reject promise to handleActiveSet() and return last known valve state
+          this.setValveServiceStatusActive(false);
           reject(this.currentValveState);
         });
     });